fixed cityio auth
